Update https.md
As a developer, i wanted to setup a wordpress environment on my organization server. During this work i encounter issues with some unexpected redirections by wordpress. I finally found and understood the problem that was linked to the way the request are transmitted. I make this update to help the next ones that will work on this topic.